+++ src/main/scala/ch/epfl/telegram/commands/TL.scala
@@ -1,18 +1,18 @@
-package ch.epfl.telegram
+package ch.epfl.telegram.commands
 
-import info.mukel.telegrambot4s.api._
+import com.github.nscala_time.time.Imports._
+import com.typesafe.emoji.ShortCodes.Defaults._
+import com.typesafe.emoji.ShortCodes.Implicits._
 import info.mukel.telegrambot4s.Implicits._
+import info.mukel.telegrambot4s.api._
 import info.mukel.telegrambot4s.methods.{EditMessageText, ParseMode}
 import info.mukel.telegrambot4s.models.{InlineKeyboardButton, InlineKeyboardMarkup}
 import net.ruippeixotog.scalascraper.browser.JsoupBrowser
 import net.ruippeixotog.scalascraper.dsl.DSL.Extract._
 import net.ruippeixotog.scalascraper.dsl.DSL._
 import org.joda.time.{DateTime, Minutes}
-import com.typesafe.emoji.ShortCodes.Implicits._
-import com.typesafe.emoji.ShortCodes.Defaults._
 
 import scala.util.Try
-import com.github.nscala_time.time.Imports._
 
 /**
   * Add TL (transport public Lausanne) useful commands.
